correcion de codigo de int a int64_t para IDchat

+++ series/telegram_bot/02_TelegramBot_Mensaje_Automatico_con_ESP32_y_ESP8266/DemoAuto/token_ejemplo.h
@@ -1,4 +1,4 @@
 const char* ssid = "xxx";
 const char* password = "xxx";
 const String token = "xxx"; 
-const int IDchat = 1234;
+const int64_t IDchat = 1234;
